Tweet every other hour

+++ bot.py
@@ -3,6 +3,7 @@ import urllib.request
 import os
 import psycopg2
 import time
+import datetime
 
 # from credentials import CONSUMER_KEY, CONSUMER_SECRET, ACCESS_KEY, ACCESS_SECRET
 from os import environ
@@ -22,15 +23,15 @@ auth.set_access_token(ACCESS_KEY, ACCESS_SECRET)
 
 twitter_API = tweepy.API(auth)
 tweets = twitter_API.user_timeline('colohistory', count=200, include_entities=True, tweet_mode='extended')
-
-result = get_random_photo(con)
-photo_id = result["id"]
-urllib.request.urlretrieve(result['imageurl'], f"./{photo_id}.jpg")
-media = twitter_API.media_upload(f"./{photo_id}.jpg")
-summary = result["summary"][:200]
-date = result["date"].split(" ")[0]
-tweet = summary + " (" + date + ") " + result['pageurl']
-
-add_shared_photo(con, result["id"])
-twitter_API.update_status(status=tweet, media_ids=[media.media_id])
-os.remove(f"./{photo_id}.jpg")
+now = datetime.datetime.now()
+if now.hour % 2 == 0
+    result = get_random_photo(con)
+    photo_id = result["id"]
+    urllib.request.urlretrieve(result['imageurl'], f"./{photo_id}.jpg")
+    media = twitter_API.media_upload(f"./{photo_id}.jpg")
+    summary = result["summary"][:200]
+    date = result["date"].split(" ")[0]
+    tweet = summary + " (" + date + ") " + result['pageurl']
+    add_shared_photo(con, result["id"])
+    twitter_API.update_status(status=tweet, media_ids=[media.media_id])
+    os.remove(f"./{photo_id}.jpg")
